drivers: power: supply: tssilo_supercaps: Fix typo & add lost props
  POWER_SUPPLY_PROP_CHARGE_CONTROL_LIMIT (RW)
  POWER_SUPPLY_PROP_CHARGE_CONTROL_LIMIT_MAX (RO)
  POWER_SUPPLY_PROP_PRESENT (RO)

These all have valid meanings, so it isn't clear why these were
removed.

The charge-enable attribute is a more "obvious" way to do what
`POWER_SUPPLY_PROP_CHARGE_CONTROL_LIMIT` does, but will surely have to
change when this driver goes upstream. There is also precedent for
using `POWER_SUPPLY_PROP_CHARGE_CONTROL_LIMIT` while they're dithering
on creating a "charge enable" property(*).

`PRESENT` exists with a hardcoded '1' because if we don't, it won't be
possible to return 0 if we ever encounter a need to.

(*) After authoring this commit, I reviewed `power_supply.h` again and
realized `POWER_SUPPLY_PROP_CHARGE_BEHAVIOUR` may be the right fit for
our needs, though it shares with other properties the semantic problem
of "is this persistent or just until power-cycled?". It doesn't
necessarily break anything here; I'm just mentioning it for if/when
this commit is reviewed/modified further.