drivers: power: supply: tssilo_supercaps: Property revisions
Instead of `POWER_SUPPLY_PROP_CHARGE_CONTROL_LIMIT`, use
`POWER_SUPPLY_PROP_CHARGE_BEHAVIOUR`. That means no further use for
`POWER_SUPPLY_PROP_CHARGE_CONTROL_LIMIT_MAX`.

We will also get rid of the `charge_enabled` attribute. It's all on
`charge_behaviour` now!

It is a little awkward in that 0 (auto) means charge and 1 (inhibit)
means no charge. But think of it as a no charge jumper, and all should
be fine.

Except, this is not at present a persistent property.

FIXME: We still need to iron out persistence.