irqchip: irq-wizard: Remove .unmask_base
For some reason, the regmap-irqchip driver doesn't understand how to
use the `IRQ_MASK_CLR` register to unmask interrupts. That is because
it lacks the notion of a mask read register being different than
either of set/clear registers. Reaching around and unmasking
interrupts on the Wizard results in everything else working OK, so it
is just this unmasking that is broken.

The Wizard as of 2025-06-19 supports the `IRQ_MASK_RW` register in
addition to the set/clear spaces.

Removing `.unmask_base` tells regmap-irqchip to treat `.mask_base` as
a R/W register, and that is now pointed at `IRQ_MASK_RW`.